+++ pygad/cloudy/cloudy_tables.py
@@ -30,12 +30,12 @@ def config_ion_table(redshift):
 
     #for old style Oppenheimer tables:
     nH_vals = iontable['nH_vals']
-    nH_vals = np.linspace(nH_vals[0],
-                          nH_vals[0] + nH_vals[1]*(nH_vals[2]-1),
+    nH_vals = np.linspace(float(nH_vals[0]),
+                          float(nH_vals[0] + nH_vals[1]*(nH_vals[2]-1)),
                           nH_vals[2])
     T_vals = iontable['T_vals']
-    T_vals = np.linspace(T_vals[0],
-                         T_vals[0] + T_vals[1]*(T_vals[2]-1),
+    T_vals = np.linspace(float(T_vals[0]),
+                         float(T_vals[0] + T_vals[1]*(T_vals[2]-1)),
                          T_vals[2])
 
     iontbl = IonisationTable(redshift,
